requirements.txt
kafka-python==0.9.5
six==1.10.0
wheel==0.24.0
